change in admin crud book,author,trans path

+++ src/main/java/com/rest_api/fs14backend/genre/GenreServiceImpl.java
@@ -1,5 +1,12 @@
 package com.rest_api.fs14backend.genre;
 
+import com.rest_api.fs14backend.author.Author;
+import com.rest_api.fs14backend.book.Book;
+import com.rest_api.fs14backend.book.BookRepository;
+import com.rest_api.fs14backend.exceptions.Genre.GenreCannotBeDeletedException;
+import com.rest_api.fs14backend.exceptions.Genre.GenreNotFoundException;
+import com.rest_api.fs14backend.exceptions.author.AuthorCannotBeDeletedException;
+import com.rest_api.fs14backend.exceptions.author.AuthorNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -11,6 +18,8 @@ import java.util.UUID;
 public class GenreServiceImpl implements GenreService {
     @Autowired
     private GenreRepository genreRepository;
+    @Autowired
+    private BookRepository bookRepository;
 
     @Override
     public List<Genre> findAll() {
@@ -41,7 +50,15 @@ public class GenreServiceImpl implements GenreService {
 
     @Override
     public void deleteOne(UUID genreId) {
-        genreRepository.deleteById(genreId);
+        Genre foundGenre =  findOneById(genreId);
+        if(foundGenre != null){
+            List<Book> foundBooks = bookRepository.findAllByAuthorId(foundGenre.getId());
+            if(foundBooks.size() > 0){
+                throw new GenreCannotBeDeletedException();
+            }
+            genreRepository.deleteById(genreId);
+        } else {
+            throw new GenreNotFoundException();
+        }
     }
-
 }
